+++ changelog.md
@@ -1,6 +1,7 @@
 ### UNRELEASED
 ### **Bruschetta**
 ---
+* LLT-3113: Add upnp endpoint provider
 * LLT-3634: `stun_plaintext_port` defaults to 0 when not provided in config
 * LLT-3603: Add support for `allow_peer_send_files` to meshconfig
 * LLT-3668: Stop Nurse before Derp